requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.2.2
chardet==3.0.4
frozenlist==1.4.1
googletrans-py39==4.0.2
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.6.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==2.10
inflect==7.3.1
magic-filter==1.0.12
more-itertools==10.4.0
multidict==6.0.5
mysql-connector-python==8.4.0
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.1
python-telegram-bot==21.1.1
pytz==2024.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==2.0.32
typeguard==4.3.0
typing_extensions==4.11.0
tzlocal==5.2
yarl==1.9.4
